+++ app/page/MainPage/Main .module.css
@@ -1,0 +1,3 @@
+.MainImage {
+  text-aligin: center;
+}
