--- gib/app/page/MainPage/Main.tsx
@@ -1,18 +1,0 @@
-import React from "react";
-
-const Main = () => {
-  return (
-    <>
-      <div className="image"></div>
-      <article>
-        <div className="image"></div>
-        <p>
-          当社はデザイナーによる完全受注生産の為、１点１点手作業で製作いたします。
-          <br></br>by GOSHI INOUE
-        </p>
-      </article>
-    </>
-  );
-};
-
-export default Main;
